// src/generated/graphql.tsx
export const SearchRepositoriesDocument = gql`
    query SearchRepositories($query: String!, $first: Int!, $after: String) {
  search(type: REPOSITORY, query: $query, first: $first, after: $after) {
    pageInfo {
      endCursor
    }
    repositoryCount
    nodes {
      ... on Repository {
        id
        name
        stargazerCount
        forkCount
      }
    }
  }
}
    `;